[SC-7718] Common Parts from AppContext for Aave Contexts
