+++ features/aave/AaveContext.ts
@@ -1,28 +1,16 @@
-import BigNumber from 'bignumber.js'
-import { GraphQLClient } from 'graphql-request'
 import { memoize } from 'lodash'
 import moment from 'moment'
 import { curry } from 'ramda'
 import { Observable, of } from 'rxjs'
-import { distinctUntilKeyChanged, map, shareReplay, switchMap } from 'rxjs/operators'
+import { switchMap } from 'rxjs/operators'
 
 import { getAaveV2AssetsPrices } from '../../blockchain/aave'
-import { getChainlinkOraclePrice } from '../../blockchain/calls/chainlink/chainlinkPriceOracle'
 import { observe } from '../../blockchain/calls/observe'
 import { TokenBalances } from '../../blockchain/tokens'
-import { UserDpmAccount } from '../../blockchain/userDpmProxies'
 import { AppContext } from '../../components/AppContext'
 import { LendingProtocol } from '../../lendingProtocols'
 import { prepareAaveTotalValueLocked$ } from '../../lendingProtocols/aave-v2/pipelines'
-import { getAllowanceStateMachine } from '../stateMachines/allowance'
-import {
-  getCreateDPMAccountTransactionMachine,
-  getDPMAccountStateMachine,
-} from '../stateMachines/dpmAccount'
-import { getGasEstimation$, getOpenProxyStateMachine } from '../stateMachines/proxy/pipelines'
-import { transactionContextService } from '../stateMachines/transaction'
 import { getAaveStEthYield } from './common'
-import { getAvailableDPMProxy$ } from './common/services/getAvailableDPMProxy'
 import {
   getAdjustAaveParametersMachine,
   getCloseAaveParametersMachine,
@@ -30,8 +18,7 @@ import {
   getOpenAaveParametersMachine,
 } from './common/services/getParametersMachines'
 import { getStrategyInfo$ } from './common/services/getStrategyInfo'
-import { getOperationExecutorTransactionMachine } from './common/services/getTransactionMachine'
-import { getProxiesRelatedWithPosition$ } from './helpers/getProxiesRelatedWithPosition'
+import { getCommonPartsFromAppContext } from './getCommonPartsFromAppContext'
 import {
   getManageAaveStateMachine,
   getManageAaveV2PositionStateMachineServices,
@@ -39,26 +26,35 @@ import {
 import { getOpenAaveStateMachine, getOpenAaveV2PositionStateMachineServices } from './open/services'
 import { getAaveSupportedTokenBalances$ } from './services/getAaveSupportedTokenBalances'
 import { getSupportedTokens } from './strategyConfig'
-import { PositionId } from './types'
 
-export function setupAaveV2Context({
-  userSettings$,
-  connectedContext$,
-  proxyAddress$,
-  txHelpers$,
-  gasPrice$,
-  daiEthTokenPrice$,
-  balance$,
-  onEveryBlock$,
-  context$,
-  tokenPriceUSD$,
-  allowance$,
-  userDpmProxies$,
-  userDpmProxy$,
-  proxyConsumed$,
-  strategyConfig$,
-  protocols,
-}: AppContext) {
+export function setupAaveV2Context(appContext: AppContext) {
+  const {
+    userSettings$,
+    txHelpers$,
+    balance$,
+    onEveryBlock$,
+    context$,
+    tokenPriceUSD$,
+    proxyConsumed$,
+    strategyConfig$,
+    protocols,
+  } = appContext
+
+  const {
+    allowanceForAccount$,
+    allowanceStateMachine,
+    dpmAccountStateMachine,
+    gasEstimation$,
+    operationExecutorTransactionMachine,
+    proxyForAccount$,
+    proxyStateMachine,
+    proxiesRelatedWithPosition$,
+    unconsumedDpmProxyForConnectedAccount$,
+    disconnectedGraphQLClient$,
+    chainlinkUSDCUSDOraclePrice$,
+    chainLinkETHUSDOraclePrice$,
+  } = getCommonPartsFromAppContext(appContext)
+
   const {
     aaveUserAccountData$,
     aaveAvailableLiquidityInUSDC$,
@@ -69,46 +65,16 @@ export function setupAaveV2Context({
     aaveOracleAssetPriceData$,
     getAaveReserveData$,
   } = protocols[LendingProtocol.AaveV2]
-  const chainlinkUSDCUSDOraclePrice$ = memoize(
-    observe(onEveryBlock$, context$, getChainlinkOraclePrice('USDCUSD'), () => 'true'),
-  )
 
-  const chainLinkETHUSDOraclePrice$ = memoize(
-    observe(onEveryBlock$, context$, getChainlinkOraclePrice('ETHUSD'), () => 'true'),
-  )
-
-  const contextForAddress$ = connectedContext$.pipe(
-    distinctUntilKeyChanged('account'),
-    shareReplay(1),
-  )
-  const disconnectedGraphQLClient$ = context$.pipe(
-    distinctUntilKeyChanged('cacheApi'),
-    map(({ cacheApi }) => new GraphQLClient(cacheApi)),
-  )
   const aaveSthEthYieldsQuery = memoize(
     curry(getAaveStEthYield)(disconnectedGraphQLClient$, moment()),
     (riskRatio, fields) => JSON.stringify({ fields, riskRatio: riskRatio.multiple.toString() }),
-  )
-
-  const gasEstimation$ = curry(getGasEstimation$)(gasPrice$, daiEthTokenPrice$)
-  const proxyForAccount$: Observable<string | undefined> = contextForAddress$.pipe(
-    switchMap(({ account }) => proxyAddress$(account)),
-  )
-
-  const proxiesRelatedWithPosition$ = memoize(
-    curry(getProxiesRelatedWithPosition$)(proxyAddress$, userDpmProxy$),
-    (positionId: PositionId) => `${positionId.walletAddress}-${positionId.vaultId}`,
   )
 
   const earnCollateralsReserveData = {
     STETH: aaveReserveConfigurationData$({ token: 'STETH' }),
   } as Record<string, ReturnType<typeof aaveReserveConfigurationData$>>
 
-  const allowanceForAccount$: (token: string, spender: string) => Observable<BigNumber> = memoize(
-    (token: string, spender: string) =>
-      contextForAddress$.pipe(switchMap(({ account }) => allowance$(token, account, spender))),
-    (token, spender) => `${token}-${spender}`,
-  )
   const aaveSupportedTokenBalances$ = memoize(
     curry(getAaveSupportedTokenBalances$)(
       balance$,
@@ -133,48 +99,6 @@ export function setupAaveV2Context({
   const closeAaveParameters = getCloseAaveParametersMachine(txHelpers$, gasEstimation$)
   const adjustAaveParameters = getAdjustAaveParametersMachine(txHelpers$, gasEstimation$)
   const depositBorrowAaveMachine = getDepositBorrowAaveMachine(txHelpers$, gasEstimation$)
-
-  const commonTransactionServices = transactionContextService(context$)
-
-  const allowanceStateMachine = getAllowanceStateMachine(
-    txHelpers$,
-    connectedContext$,
-    commonTransactionServices,
-  )
-
-  const proxyStateMachine = getOpenProxyStateMachine(
-    contextForAddress$,
-    txHelpers$,
-    proxyForAccount$,
-    gasEstimation$,
-  )
-
-  const dpmAccountTransactionMachine = getCreateDPMAccountTransactionMachine(
-    txHelpers$,
-    connectedContext$,
-    commonTransactionServices,
-  )
-  const dpmAccountStateMachine = getDPMAccountStateMachine(
-    txHelpers$,
-    gasEstimation$,
-    dpmAccountTransactionMachine,
-  )
-
-  const operationExecutorTransactionMachine = curry(getOperationExecutorTransactionMachine)(
-    txHelpers$,
-    contextForAddress$,
-    commonTransactionServices,
-  )
-
-  const getAvailableDPMProxy: (
-    walletAddress: string,
-  ) => Observable<UserDpmAccount | undefined> = memoize(
-    curry(getAvailableDPMProxy$)(userDpmProxies$, proxyConsumed$),
-  )
-
-  const unconsumedDpmProxyForConnectedAccount$ = contextForAddress$.pipe(
-    switchMap(({ account }) => getAvailableDPMProxy(account)),
-  )
 
   const openAaveStateMachineServices = getOpenAaveV2PositionStateMachineServices(
     context$,
